core/events: Continue splitting events
