+++ src/natwm/natwm.c
@@ -15,7 +15,7 @@
 #include <common/map.h>
 #include <common/util.h>
 #include <core/config/config.h>
-#include <core/event.h>
+#include <core/events/event.h>
 #include <core/ewmh.h>
 #include <core/monitor.h>
 #include <core/state.h>
